requirements.txt
# Dependencias para automation-hub

# HTTP requests
requests>=2.31.0

# Google Authentication
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.116.0

# Date/time handling
python-dateutil>=2.8.2
pytz>=2024.1

# Supabase client
supabase>=2.6.0

# Environment variables
python-dotenv>=1.0.1

# Dashboard server
flask>=3.0.0
flask-cors>=4.0.0

# Meta (Facebook) Marketing API
facebook-business>=19.0.0
